Moved IC logic to a separate file, renamed wrapper
The Rodan wrapper is now in "wrapper.py"
